feat: move finality monitoring into frameowrk
